+++ changelog.md
@@ -1,8 +1,12 @@
 # Changelog
 
+## 0.2.0
+- fix route_prefix validation
+- update layouts to show correct categories
+- add return local view i exists
+
 ## 0.1.4
 - add previewRoute
-
 
 ## 0.1.3
 - fix title error when not available